refactor: add active tag
